netfilter: Fix multicast flow is not applied SNAT correctly
When bridge_nf_call_iptable is enabled, multicast flow
can reach nf_nat_inet_fn twice for SNAT operation, the
first time, it floods/snoops all ports of the bridge,
it won't hit nat rule in the do_chain, so
no nat operation required and finalized the conntrack,
and the second time, it reach nf_nat_inet_fn for forwarding
bridge to an upstream port, no do_chain is performed and
using the first time's result.

This is not the expected behavior. If the multicast is
forwarding between ports of a bridge, accept it directly.

Change-Id: I3dd3be2bd718051292a46485a1863cf127af4866
Signed-off-by: Ken Zhu <guigenz@codeaurora.org>
Signed-off-by: Pavithra R <quic_pavir@quicinc.com>