OpenWrt: 683-of_net-add-mac-address-to-of-tree.patch
of/of_net: write back netdev MAC-address to device-tree

The label-mac logic relies on the mac-address property of a netdev
devices of-node. However, the mac address can also be stored as a
different property or read from e.g. an mtd device.

Create this node when reading a mac-address from OF if it does not
already exist and copy the mac-address used for the device to this
property. This way, the MAC address can be accessed using procfs.

Change-Id: I3d4731839ffe2ec83cc7e237ccce29c8f46ce718
Signed-off-by: Ram Chandra Jangir <quic_rjangir@quicinc.com>
(cherry picked from commit 40b96cc9d18f260af9bec2eacd196e0a8b89623b)