net: Add extra priv_flag support in netdevice
This patch adds a flag in struct net_device which
is an extension to priv_flags and is used to
check the type of device.

Change-Id: Ibb5a3f0c706e1b169b848aed7004aa4ea6d57069
Signed-off-by: Pavithra R <pavir@codeaurora.org>

Add GRE std support in device tree and priv_flags

Added dts support and priv flags support for GRE and
to check the type of device

Change-Id: Ib05a32e9e5458c724c182187362116338d767169
Signed-off-by: ratheesh kannoth <rkannoth@codeaurora.org>
Signed-off-by: Pavithra R <pavir@codeaurora.org>

drivers: ifb: Add flag to identify the IFB interface.

Add IFB flag in netdevice's ext priv flags and define an API
to let other modules find out if the netdev is IFB or not.

Change-Id: I6dbdae521240b4a991069c2330d893ff38aec68d
Signed-off-by: Manish Verma <maniverm@codeaurora.org>
Signed-off-by: pavir <pavir@codeaurora.org>

net: add IFF_EXT_MAPT flag in net_device priv_flags_ext

Add IFF_EXT_MAPT flag in priv_flags_ext for MAP-T device.

Change-Id: I8aa1715c4b9c62b88eee1b40412cf4e734b2a337
Signed-off-by: Suruchi Agarwal <quic_suruchia@quicinc.com>
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>