remoteproc: qcom: Add NOTIFY_FATAL event type to SSR subdevice
Currently the SSR subdevice notifies the client driver on crash of the
rproc from the recovery workqueue using the BEFORE_SHUTDOWN event.
However the client driver might be interested to know that the device
has crashed as soon as possible atomically.

Add NOTIFY_FATAL event to SSR subdevice to atomically notify rproc has
crashed to the client driver.

Change-Id: Ia6b97007e34fc3898f14c90ca06aa44565484a4e
Signed-off-by: Sridharan S N <quic_sridsn@quicinc.com>