pcie: qcom: ipq9574: Forcing PCIE2/3 to single lane
For lane switching, as num-lanes property is not
supported in Alder. This patch enables the support
for switching PCIE2/3 to single lane. The function
is added in PCI_FINAL_FIXUP and can be enabled from
the DTS node by defining "force_to_single_lane" property.

NOTE: Do not change num-lanes to 1 for PCIE2/3.
Instead, "force_to_single_lane" property can be defined
in DTS for PCIE2/3 to switch the corresponding node to
single lane.

Change-Id: I14d29f436416868661d103579f71898748078d22
Signed-off-by: Sridharan S N <quic_sridsn@quicinc.com>