mtd: ubi: introduce pre-removal notification for UBI volumes
Introduce a new notification type UBI_VOLUME_SHUTDOWN to inform users
that a volume is just about to be removed.
This is needed because users (such as the NVMEM subsystem) expect that
at the time their removal function is called, the parenting device is
still available (for removal of sysfs nodes, for example, in case of
NVMEM which otherwise WARNs on volume removal).

Signed-off-by: Daniel Golle <daniel@makrotopia.org>
Signed-off-by: Richard Weinberger <richard@nod.at>
(cherry picked from commit 7e84c961b2eb062d2f47037dcca52dcd1d3615b5)

Change-Id: I2ee2bbfd93736bd7084d9678cac931f231582389
Signed-off-by: Saahil Tomar <quic_saahtoma@quicinc.com>