pcie: qcom: ipq9574: Enable Virtual channel for QCN9224
This patch enables the VC1 if the end point is QCN9224.
Fucntion added in the PCI_FINAL_FIXUP and feature can be
enabled from the DTS node using "enable_virtual_channel"
flag. This will enable the VC1 on both IPQ9574 and QCN9224.

This is a snapshot of the commit
3777826ba4d3fbc0294eb8028f0e7bd7853cafe6
from win.linuxopenwrt.2.1 branch and
oss/kernel/linux-ipq-5.4 project.

Change-Id: Ie0cbc64697bc1d00e0ecc4a9e4caeba1953c0888
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>
Signed-off-by: Sridharan S N <quic_sridsn@quicinc.com>