watchdog: qcom: add support to trigger the NS WDT bite on kernel panic
Set the priority of the WDT restart handler to 255, so that it takes
precedence over other handlers. With this approach, all the system
reset will be done via watchdog. To make sure, WDT bite is triggered
only on panic, register a panic notifier and update that system is in
panic path. Based on that condition, in qcom_wdt_restart trigger the
watchdog bite.

Gaurd these changes under separate Kconfig so that it can be enabled
only for the required targets.

Change-Id: If5ea7208bc075694201e135c90af592aac1a2d48
Signed-off-by: Kathiravan Thirumoorthy <quic_kathirav@quicinc.com>