nvmem: core: Print error on wrong bits DT property
The algorithms in nvmem core are built with the constraint that
bit_offset < 8. If bit_offset is greater the results are wrong. Print an
error if the devicetree 'bits' property is outside of the valid range
and abort parsing.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Link: https://lore.kernel.org/r/20240224114516.86365-12-srinivas.kandagatla@linaro.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit def3173d4f17b37cecbd74d7c269a080b0b01598)

Change-Id: If2fd1f292f9c2d7e1107994c8c74be42367a75ca
Signed-off-by: Saahil Tomar <quic_saahtoma@quicinc.com>