cfg80211: Reset valid_link bitmap after the link is deleted
In 6.6 kernel, cfg80211_remove_link is called only for the links
present as part of wdev->valid_links. The other links as part of MLD
group but not part of connection are not deleted.

Reset wdev->valid_links bitmap after calling the rdev_del_intf_link
API so that the deletion of other links are taken care in the
driver based on the current valid_links count.

Change-Id: I7323d0008fe6347bc2e424887cc7f6fd9213e0cb
Signed-off-by: Abishek Ganapathy <quic_abishekg@quicinc.com>