coresight: tmc: handle the error cases properly
Probe of TMC ETR is deferred until the rproc device is available. In
such cases, we need to unregister the panic notifier, else system will
crash due to the UAF scenario.

Only for certain targets, we need to abort the TMC and write the markers
during SSR event. So when the qcom,rproc property is not available in
tmc node, skip the SSR handler registration, instead of returning the
error.

Change-Id: I7179573173b97113c184ca4ccd512a084ae3d827
Fixes: 2cba54300cd7 ("coresight: tmc-etr: add coresight ssr abort support")
Signed-off-by: Kathiravan Thirumoorthy <quic_kathirav@quicinc.com>