Updating 683-of_net-add-mac-address-to-of-tree.patch
[PATCH] of/of_net: write back netdev MAC-address to device-tree

The label-mac logic relies on the mac-address property of a netdev
devices of-node. However, the mac address can also be stored as a
different property or read from e.g. an mtd device.

Create this node when reading a mac-address from OF if it does not
already exist and copy the mac-address used for the device to this
property. This way, the MAC address can be accessed using procfs.

Change-Id: Id645fec195435047186b92cbdbf0b946cfd15491
Signed-off-by: Abhinav Kishan <quic_apothara@quicinc.com>