phy: qcom-qusb2: Set device driver_data early in probe
In qusb2_phy_probe, qphy is set as driver_data to the device towards the
end of the probe. pm_runtime ops registered for this device is enabled
before the driver_data is set but in the pm_ops, dev_get_drvdata is used
to get qphy and dereferenced without NULL checks. This causes a NULL
pointer dereference if pm_ops are invoked before the probe is completed.

Move setting driver_data to earlier in the probe as soon as qphy is
allocated.

Change-Id: I848521140c4524196ff4949c8a33d93c000d33c8
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>