cfg80211: Fix the scan resource busy issue
When mldX / athX is dev_close from netifd, notifier call
does the abort of pending scan_req if any. After aborting,
though there is no requirement of sending scan_msg to
application, cfg is preparing the scan msg and keeping it
in kernel forever as driver can not send the scan_done
since netdev is in down state. This causes the further
scan requests from application to fail with Resource busy
(-16/-EBUSY) error, after the netdev is brought UP back.

To fix this scenario, abort the scan on sta interface as
part of cfg80211_remove_links. This informs driver that
the interface going down, and no need to continue scan
further. Upon this,

1. If driver sends ack to abort scan & if its completed
first before NETDEV_DOWN, work queue itself sends scan
abort status message to application.
2. If driver could not cancel the scan immediately & could
not ack back to cfg, NETDEV_DOWN completes & cfg itself
aborts scan & consume the same as the netdev is already in
down state and work queue will not be scheduled for this
scan request further
3. If driver itself aborts the scan & initiates dev_close
of the interface, work_queue is scheduled first, and if
NETDEV_DOWN completed immediately, the API waits for the
work queue to post the rdev->scan_msg to application

Change-Id: Ic2124590d79dcca48b3251b9ed9f8d026133bd0f
Signed-off-by: Shwetha G K <quic_kshwetha@quicinc.com>