cfg80211: Don't reset ssid len and nlportid after a link removal.
During a link removal from a MLD, stop ap is called
where nlportid is getting reset to 0. Which is wrong
as still the partner links and MLD will be up and running.
It causes the next consequent association to fail by rejecting M2
packet. This change fixes by resetting the nlportid only during
the last link removal/if all links are already
removed.

Same operation for ssid len as well.

Change-Id: I395a657076e548b3bbcd6d845e36f57731e9d251
Signed-off-by: Mohan Prabu <quic_mprabub@quicinc.com>