nl80211/cfg80211: Use BUG_ON instead of WARN_ON for invalid chandef
When the channel definition is invalid, kernel gives a warning
and prints a stack trace. This warning is not of much help to identify
the reason why the chandef is invalid. For better debugging and to
collect dumps for knowing the complete state of the system, replace
WARN_ON with BUG_ON to force crash the system when the chendef is
invalid.

Change-Id: Ie8718d63e30819a72454d7949de5b943728390f3
Signed-off-by: Thirusenthil Kumaran J <quic_thirusen@quicinc.com>