watchdog: qcom: add support to read the restart reason from IMEM
When the system boots up after a watchdog reset, the EXPIRED_STATUS bit
in the WDT_STS register is cleared. To identify if the system was
restarted due to WDT expiry, XBL update the information in the IMEM
region. Update the driver to read the restart reason from IMEM and populate the bootstatus accordingly.

With the CONFIG_WATCHDOG_SYSFS enabled, user can extract the information
as below:

cat
/sys/devices/platform/soc@0/f410000.watchdog/watchdog/watchdog0/bootstatus
32

For backward compatibility, keep the EXPIRED_STATUS bit check. Add a new
function qcom_wdt_get_restart_reason() to read the restart reason from
IMEM.

Change-Id: I95d19e0a70705a8d0e2f7df5804b527d40962f37
Signed-off-by: Kathiravan Thirumoorthy <quic_kathirav@quicinc.com>