thermal: core: Limit upper cooling state to max supported state
While binding a cooling device to a thermal zone, if the requested
upper cooling state is greater than the max supported state it is
considered as an error. In some scenarios, the cpufreq opp table can
change dynamically based on OTP fuse during bootup. Due to this, the
MIN and MAX cooling states hardcoded in the device tree node can be
greater than the max supported indices in the cpufreq opp table and
thus become INVALID. In such cases, instead of considering it as an
error, we can limit the upper state to the max supported state. By
doing so, if the lower state exceeds the upper state then it will
also be limited to the upper state.

Change-Id: I9832611f6f21fa82efa16dad4b12bda27019eb7c
Signed-off-by: Hariharan K <harihk@qti.qualcomm.com>