EHT: changes to add EHT atomic notifier callback in Linux
This Patch invokes an atomic notifier callback function when there is a meaningful change of state to the EHT(Enhanced-Host-Tracking) database for a Multicast listener port (Wifi port).

During this notification call, linux will send out various information to the Wifi driver for it to help perform MCUC (Multicast-to-Unicast) conversion for various hosts.

Wifi driver will listen to this notification to update it's internal database on the fly and keep track on the active MCAST listeners to make sure it sends out MCAST data to those relevant hosts only.

Change-Id: I188336f0c720cd7826ddfbe38d1539f29dbc177c
Signed-off-by: Suhas N Bhargav <sbhargav@qti.qualcomm.com>
Signed-off-by: KRITHI D SHETTY <kdshetty@qti.qualcomm.com>