drivers: pci: qcom: Disable clocks directly to reset RC after early cal
Currently, for early cal enabled case, as part of RC init sequence added
via I7ab6fc70579e15774bc5717d6b3f63435fe71eef, RC clocks needs to be
disabled as it was enabled by u-boot, but clock framework ref counts
does not allow to disable clocks directly. So clk_enable was invoked
before disabling the clocks.

However, this is causing the clocks to go to bad state and randomly
while disabling the clocks, system goes for a secure watchdog.

Add new API in clk driver to bypass the ref count checks and directly
disable the hw_clk and use this API in the PCIe reset sequence for early
cal.

Change-Id: If349bc0ae8e0ab6ae2d6957288dffbbda06021cd
Signed-off-by: Vignesh Viswanathan <viswanat@qti.qualcomm.com>