MDB: changes to add MDB atomic notifier callback in Linux
This Patch invokes an atomic notifier callback in (br_mdb_notify) function when there is a meaningful change of state to a port-group (PG) for a Multicast listener port.

ECM listens to this notification to update it's interface heirarchies on the fly to make sure it's rule is always up-to-date.

Change-Id: Ib2905b0338750cd2c94899b6370247e97f4964dc
Signed-off-by: Suhas N Bhargav <sbhargav@qti.qualcomm.com>
Signed-off-by: KRITHI D SHETTY <kdshetty@qti.qualcomm.com>