net: Add chan to freq formula for new channels
Add channel to frequency conversion formula for
2.5MHZ step size and Edge channels in ieee80211_channel_to_frequency()
API.This API is invoked by the driver when the frame is passed on
to cfgi80211 layer as a part of scan.

cfg80211: Allow 2.4GHz Edge channels with 5/10 Mhz BW support
The AP is not able to come up on 2.4GHz newly added Edge channels
with following center frequencies 2402,2404,2407,2409,
2492,2494,2497 and 2499 Mhz

Since 20 MHz is the minimum bandwidth configured in freq_reg_info,
cfg80211_does_bw_fit_range fails to return true for the above frequencies.

To fix this issue, in cfg80211_does_bw_fit_range, if the center_freq
falls in the 2.4GHz new channels range, then do not modify
the start_freq and end_freq

Change-Id: Ib3f7c80513b7521ec74d1e7b02d85f1d55f8f4df
Signed-off-by: Naresh Pirati <npirati@qti.qualcomm.com>