net: allow driver to select valid 320 MHz center frequency on 6 GHz when ccfs is not configured
When an AP is configured on 6 GHz with 320 MHz bandwidth without
a center frequency segment (ccfs) configuration, the hostapd script
currently selects the nearest center_freq1. This can result in
choosing a disabled channel as center_freq1 or a center_freq1 which
belongs to a disabled channel block, causing the kernel to reject
the start_ap request. Consequently, the AP fails to come up on the
given primary channel. The issue occurs when the 320 MHz block of the
selcted center_freq1 contains disabled channels.

To address this issue, modify the kernel logic to allow the driver
to select a valid 320 MHz center frequency when ccfs is not configured.
Specifically:
  - If bandwidth is 320 MHz and band is 6 GHz, set center_freq1 to 0.
  - Pass this value to the driver, enabling it to choose an valid
    center frequency for the configuration.

Change-Id: Ibec7dc081fc5e02ae9634193c5491f243c55c40a
Signed-off-by: Lokitha Thangavel <lthangav@qti.qualcomm.com>